Made DeleteWord,SaveWord more correct in order of duplicates. Bug: bot is too sensetive for CapsLock in words in commands

+++ src/main/java/com/alexey/commands/DeleteWord.java
@@ -1,5 +1,6 @@
 package com.alexey.commands;
 
+import com.alexey.repository.WordRepository;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import com.alexey.models.Word;
 import com.alexey.service.WordService;
@@ -12,31 +13,40 @@ import com.alexey.Bot;
 public class DeleteWord {
     @Autowired
     private WordService wordService;
+    @Autowired
+    private WordRepository wordRepository;
     private Bot bot;
 
 public SendMessage DeleteWord(String message, Long ChatId) throws TelegramApiException {
 
     String[] words = message.split(" ", 3);
     SendMessage sendMessage = new SendMessage();
-    sendMessage.setChatId(ChatId);
     String word = words[1];
     String translation = words[2];
-    if(words.length == 3) { // Идет проверка, что клиент разделил на 3 части запрос
+    if(wordRepository.existsByWordAndTranslation(word, translation)) {
+        if(words.length == 3) { // Идет проверка, что клиент разделил на 3 части запрос
 
-        wordService.DeleteWord(word,translation); // Удаляет слово по конструкции "слово-перевод"
-        sendMessage.setChatId(ChatId);
-        sendMessage.setText("Пара: "+word+"-"+translation+" удалена!");
-        return sendMessage;
+            wordService.DeleteWord(word,translation); // Удаляет слово по конструкции "слово-перевод"
+            sendMessage.setChatId(ChatId);
+            sendMessage.setText("Пара: "+word+"-"+translation+" удалена!");
+            return sendMessage;
 
 
+        }
+        else{ // Клиент не разделил на 3 части запрос
+
+            sendMessage.setChatId(ChatId);
+            sendMessage.setText("Введите пару для удаления /delete dog - собака");
+            return sendMessage;
+
+        }
     }
-    else{ // Клиент не разделил на 3 части запрос
-
+    else{
         sendMessage.setChatId(ChatId);
-        sendMessage.setText("Введите пару для удаления /delete dog - собака");
+        sendMessage.setText("Такой пары слово-перевод не существует! Возможно, она была удалена ранее");
         return sendMessage;
-
     }
+
